Made item redelivery function properly. -Items get sent properly in async. -Potions and traps are not sent in async. -All items are re-sent on new game, except for traps and potions.
I ended up removing the ShouldDeliverItem line from RECV_MSG.ReceivedItems because I couldn't figure out a good way to fit it in properly, and I don't really understand it still. But it seems to be working without it?

Added an is_redeliverable_item table to item_utils. Purpose is to just clearly define whether an item should be redelivered. There was discussion before on whether respawns should be redelivered, so feel free to change that to false. Maybe we'll make it an option later.