Orbs now correctly spawn! All it took was 33 separate orb files, and would probably take a single small block of code for someone more knowledgeable with lua! But maybe Noita is just frustrating like that, idk.
Added some more items to the item list, mostly silly stuff. One less silly thing is adding the map perk to the item list, with the intention that it only gets shuffled in if Toveri is one of the checks. Note to self: make sure it doesn't get shuffled and placed on Toveri.

Added a couple of debug wands, as well as the unlimited spells wand to the debug perks.
One of the debug wands is just a digging wand straight out of Noita's files. The other is a wand pulled from the wiki to travel between parallel worlds -- to make testing stuff regarding them easier.

This is going to be a draft pull request since shops are DEFINITELY broken in this one.