Added Design Intent and Notes and Quirks section
There's some things, such as potion redelivery being disabled on spawn and gold not getting redelivered, that may look like bugs or issues. This section is to explain that they are intentional design choices, so that people don't submit bug reports for things that aren't bugs, and to give a little more insight as to how the world is laid out.

For the design intent section, I figured it would make sense to contextualize why Noita works well in Archipelago.