Crash fix on pedestal item error
Gave pedestal items name and description in the event that they don't properly get them (in the event that your datapackage isn't there, or if it wasn't there fast enough). Going to follow this up with something to make the pedestal items or hidden chests spawn normally in the event the datapackage didn't show up, it's just good to have a back up plan.